+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 ### Fix
 - Fix safe area when setting the child VCs directly
 
-### Removes
+### Removed
 - Drop support of iOS 11, 12 and 13
+- Remove workaround property `ad_isTranslucent`, use `UINavigationBar.appearance().isTranslucent` instead.
 
 ## [1.0.4]